+++ pkg/processor/sumologicschemaprocessor/aggregate_attributes_processor.go
@@ -15,6 +15,7 @@
 package sumologicschemaprocessor
 
 import (
+	"fmt"
 	"regexp"
 	"strings"
 
@@ -71,21 +72,63 @@ func pairToAggregation(pair *aggregationPair) (*aggregation, error) {
 
 func (proc *aggregateAttributesProcessor) processLogs(logs plog.Logs) error {
 	for i := 0; i < logs.ResourceLogs().Len(); i++ {
-		proc.aggregateAttributes(logs.ResourceLogs().At(i).Resource().Attributes())
+		resourceLogs := logs.ResourceLogs().At(i)
+		err := proc.processAttributes(resourceLogs.Resource().Attributes())
+		if err != nil {
+			return err
+		}
+
+		for j := 0; j < resourceLogs.ScopeLogs().Len(); j++ {
+			scopeLogs := resourceLogs.ScopeLogs().At(j)
+			for k := 0; k < scopeLogs.LogRecords().Len(); k++ {
+				err := proc.processAttributes(scopeLogs.LogRecords().At(k).Attributes())
+				if err != nil {
+					return err
+				}
+			}
+		}
 	}
 	return nil
 }
 
 func (proc *aggregateAttributesProcessor) processMetrics(metrics pmetric.Metrics) error {
 	for i := 0; i < metrics.ResourceMetrics().Len(); i++ {
-		proc.aggregateAttributes(metrics.ResourceMetrics().At(i).Resource().Attributes())
+		resourceMetrics := metrics.ResourceMetrics().At(i)
+		err := proc.processAttributes(resourceMetrics.Resource().Attributes())
+		if err != nil {
+			return err
+		}
+
+		for j := 0; j < resourceMetrics.ScopeMetrics().Len(); j++ {
+			scopeMetrics := resourceMetrics.ScopeMetrics().At(j)
+			for k := 0; k < scopeMetrics.Metrics().Len(); k++ {
+				err := processMetricLevelAttributes(proc, scopeMetrics.Metrics().At(k))
+				if err != nil {
+					return err
+				}
+			}
+		}
 	}
 	return nil
 }
 
 func (proc *aggregateAttributesProcessor) processTraces(traces ptrace.Traces) error {
 	for i := 0; i < traces.ResourceSpans().Len(); i++ {
-		proc.aggregateAttributes(traces.ResourceSpans().At(i).Resource().Attributes())
+		resourceSpans := traces.ResourceSpans().At(i)
+		err := proc.processAttributes(resourceSpans.Resource().Attributes())
+		if err != nil {
+			return err
+		}
+
+		for j := 0; j < resourceSpans.ScopeSpans().Len(); j++ {
+			scopeSpans := resourceSpans.ScopeSpans().At(j)
+			for k := 0; k < scopeSpans.Spans().Len(); k++ {
+				err := proc.processAttributes(scopeSpans.Spans().At(k).Attributes())
+				if err != nil {
+					return err
+				}
+			}
+		}
 	}
 	return nil
 }
@@ -98,6 +141,53 @@ func (*aggregateAttributesProcessor) ConfigPropertyName() string {
 	return "aggregate_attributes"
 }
 
-func (proc *aggregateAttributesProcessor) aggregateAttributes(attributes pcommon.Map) {
+func (proc *aggregateAttributesProcessor) processAttributes(attributes pcommon.Map) error {
+	for i := 0; i < len(proc.aggregations); i++ {
+		curr := proc.aggregations[i]
+		names := []string{}
+		attrs := []pcommon.Value{}
 
+		for j := 0; j < len(curr.patternRegexes); j++ {
+			regex := curr.patternRegexes[j]
+			newMap := pcommon.NewMap()
+			newMap.EnsureCapacity(attributes.Len())
+
+			attributes.Range(func(key string, value pcommon.Value) bool {
+				match := regex.FindStringSubmatch(key)
+				if match != nil {
+					// Join all substrings caught by wildcards into one string,
+					// this string will be the name of this key in the new map.
+					// TODO: Potential name conflict to resolve, eg.:
+					// pod_*_bar_* matches pod_foo_bar_baz
+					// pod2_*_bar_* matches pod2_foo_bar_baz
+					// both will be renamed to foo_baz
+					name := strings.Join(match[1:], "_")
+					names = append(names, name)
+					val := pcommon.NewValueEmpty()
+					value.CopyTo(val)
+					attrs = append(attrs, val)
+				} else {
+					value.CopyTo(newMap.PutEmpty(key))
+				}
+				return true
+			})
+			newMap.CopyTo(attributes)
+		}
+
+		if len(names) != len(attrs) {
+			return fmt.Errorf(
+				"internal error: number of values does not equal the number of keys; len(keys) = %d, len(values) = %d",
+				len(names),
+				len(attrs),
+			)
+		}
+
+		aggregated := attributes.PutEmptyMap(curr.attribute)
+
+		for j := 0; j < len(names); j++ {
+			attrs[j].CopyTo(aggregated.PutEmpty(names[j]))
+		}
+	}
+
+	return nil
 }